% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootpval.R
\name{bootpval}
\alias{bootpval}
\title{bootpval}
\usage{
bootpval(x, conf.level = 0.95, iter = 3000, mu0 = 0, test = "two")
}
\arguments{
\item{x}{numeric vector}

\item{conf.level}{confidence level}

\item{iter}{iterations}

\item{mu0}{mean}

\item{test}{type of test}
}
\value{
graph and list of pvalue, tcalc, n, x, test, and ci
}
\description{
bootpval
}
\examples{
bootpval(1:10)
}
