% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage data}
\format{
## `fire`
A data frame with 15 rows and 2 columns:
\describe{
\item{DISTANCE}{distance}
\item{DAMAGE}{damage}
...
}
}
\source{
<FIREDAM.csv>
}
\usage{
fire
}
\description{
A subset of fire damage data
Report ...
}
\keyword{datasets}
