% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMode.R
\name{getmode}
\alias{getmode}
\title{The Square Function}
\usage{
getmode(v)
}
\arguments{
\item{v}{vector of data}
}
\value{
The mode of the data.
}
\description{
The Square Function
}
\examples{
getmode(1:10)
}
