% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{mybin}
\usage{
mybin(iter = 100, n = 10, p = 0.5, main = "Binomial simulation")
}
\arguments{
\item{iter}{number of iterations}

\item{n}{size of sample}

\item{p}{probability}

\item{main}{plot title}
}
\value{
barplot and table of the proportions
}
\description{
mybin
}
\examples{
mybin(iter=100, n=10, p=0.5)
}
