% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{mycltp}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{Sample Size}

\item{iter}{Iteration}

\item{lambda}{Lambda}

\item{...}{Other parameters}
}
\value{
Graphs and w using random sample from Poisson.
}
\description{
mycltp
}
\examples{
mycltp(n=30, iter=10000, lambda=10)
}
