% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlbeta.R
\name{mymlbeta}
\alias{mymlbeta}
\title{beta distribution max liklihood}
\usage{
mymlbeta(x, alpha, beta, ...)
}
\arguments{
\item{x}{numeric vector}

\item{alpha}{alpha}

\item{beta}{beta}

\item{...}{other parameters}
}
\value{
Max liklihood graph for beta distribution
}
\description{
beta distribution max liklihood
}
\examples{
mymlbeta(x=rbeta(100,shape1=2,shape2=5),alpha=seq(1,4,length=100),beta=seq(2,8,length=100),lwd=2,labcex=1)
}
