% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve function}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean}

\item{sigma}{standard deviation}

\item{a}{probability (X <= a)}
}
\value{
chart of curve with region shaded with probability.
}
\description{
myncurve function
}
\examples{
myncurve(1, 2, 4)
}
