% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets function}
\usage{
ntickets(N = 200, gamma = 0.02, p = 0.95)
}
\arguments{
\item{N}{Number of seats on the flight}

\item{gamma}{gamma}

\item{p}{probability}
}
\value{
Discrete and continuous plots of Objective vs N tickets sold.
}
\description{
ntickets function
}
\examples{
ntickets(400, 0.01, 0.95)
}
